Initial framework and tests for Aruba::Platform operating system, ruby_platform, and ruby_engine
This can be expanded to support testing (with mocks, etc.) different operating systems.  The methods is Aruba::Platform ask FFI:Platform what the current operating system is.  In these tests, FFI::Platform is replaced by a mock (a class double) so that the answer to 'what is the operating system?' can be instrumented as needed.  The ruby_platform and ruby_engine are simply determined by the value of the global constants RUBY_PLATFORM and RUBY_ENGINE.  These are mocked by the 'stub_const()' method, so the value(s) can be changed as needed to test the system behavior.